Encapsulate Iterator implementation in one place
