+++ spinoso-string/src/impls.rs
@@ -2,19 +2,12 @@ use alloc::borrow::Cow;
 use alloc::vec::Vec;
 use core::borrow::{Borrow, BorrowMut};
 use core::fmt;
-use core::iter::FusedIterator;
 use core::ops::{Deref, DerefMut, Index, IndexMut};
 use core::slice::SliceIndex;
 #[cfg(feature = "std")]
 use std::io;
 
-use crate::{Bytes, Center, IntoIter, Iter, IterMut, String};
-
-impl<'a> AsRef<[u8]> for Iter<'a> {
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-}
+use crate::String;
 
 impl fmt::Write for String {
     #[inline]
@@ -63,233 +56,6 @@ impl io::Write for String {
     #[inline]
     fn flush(&mut self) -> io::Result<()> {
         self.buf.flush()
-    }
-}
-
-impl<'a> Iterator for Iter<'a> {
-    type Item = &'a u8;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-
-    #[inline]
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth(n)
-    }
-
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-
-    #[inline]
-    fn last(self) -> Option<Self::Item> {
-        self.0.last()
-    }
-}
-
-impl<'a> DoubleEndedIterator for Iter<'a> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.0.next_back()
-    }
-
-    #[inline]
-    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth_back(n)
-    }
-}
-
-impl<'a> FusedIterator for Iter<'a> {}
-
-impl<'a> ExactSizeIterator for Iter<'a> {}
-
-impl<'a> Iterator for IterMut<'a> {
-    type Item = &'a mut u8;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-
-    #[inline]
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth(n)
-    }
-
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-
-    #[inline]
-    fn last(self) -> Option<Self::Item> {
-        self.0.last()
-    }
-}
-
-impl<'a> DoubleEndedIterator for IterMut<'a> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.0.next_back()
-    }
-
-    #[inline]
-    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth_back(n)
-    }
-}
-
-impl<'a> FusedIterator for IterMut<'a> {}
-
-impl<'a> ExactSizeIterator for IterMut<'a> {}
-
-impl AsRef<[u8]> for IntoIter {
-    #[inline]
-    fn as_ref(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-}
-
-impl Iterator for IntoIter {
-    type Item = u8;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
-    }
-
-    #[inline]
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth(n)
-    }
-
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-
-    #[inline]
-    fn last(self) -> Option<Self::Item> {
-        self.0.last()
-    }
-}
-
-impl DoubleEndedIterator for IntoIter {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.0.next_back()
-    }
-
-    #[inline]
-    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth_back(n)
-    }
-}
-
-impl FusedIterator for IntoIter {}
-
-impl ExactSizeIterator for IntoIter {}
-
-impl<'a> Iterator for Bytes<'a> {
-    type Item = u8;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        self.0.next().copied()
-    }
-
-    #[inline]
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth(n).copied()
-    }
-
-    #[inline]
-    fn count(self) -> usize {
-        self.0.count()
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.0.size_hint()
-    }
-
-    #[inline]
-    fn last(self) -> Option<Self::Item> {
-        self.0.last().copied()
-    }
-}
-
-impl<'a> DoubleEndedIterator for Bytes<'a> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.0.next_back().copied()
-    }
-
-    #[inline]
-    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
-        self.0.nth_back(n).copied()
-    }
-}
-
-impl<'a> FusedIterator for Bytes<'a> {}
-
-impl<'a> ExactSizeIterator for Bytes<'a> {}
-
-impl<'a, 'b> Iterator for Center<'a, 'b> {
-    type Item = u8;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(&next) = self.left.next() {
-            return Some(next);
-        }
-        if let Some(next) = self.next.take() {
-            if let Some((&first, tail)) = next.split_first() {
-                self.next = Some(tail);
-                return Some(first);
-            }
-        }
-        if let Some(next) = self.s.next() {
-            if let Some((&first, tail)) = next.split_first() {
-                if !tail.is_empty() {
-                    self.next = Some(tail);
-                }
-                return Some(first);
-            }
-        }
-        self.right.next().copied()
-    }
-}
-
-impl<'a, 'b> FusedIterator for Center<'a, 'b> {}
-
-impl<'a, 'b> ExactSizeIterator for Center<'a, 'b> {}
-
-impl IntoIterator for String {
-    type Item = u8;
-    type IntoIter = IntoIter;
-
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        IntoIter(self.buf.into_iter())
     }
 }
