+++ spinoso-string/src/lib.rs
@@ -34,14 +34,13 @@ extern crate alloc;
 extern crate std;
 
 use alloc::boxed::Box;
-use alloc::vec::{self, Vec};
+use alloc::vec::Vec;
 use core::cmp::Ordering;
 use core::fmt::{self, Write};
 use core::hash::{Hash, Hasher};
-use core::iter::{Cycle, Take};
 use core::mem;
 use core::ops::Range;
-use core::slice::{self, SliceIndex};
+use core::slice::SliceIndex;
 use core::str;
 
 use bstr::{ByteSlice, ByteVec};
@@ -58,365 +57,13 @@ mod encoding;
 mod eq;
 mod impls;
 mod inspect;
+mod iter;
 
 pub use chars::Chars;
 pub use codepoints::{Codepoints, CodepointsError};
 pub use encoding::{Encoding, InvalidEncodingError};
 pub use inspect::Inspect;
-
-/// Immutable [`String`] byte slice iterator.
-///
-/// This struct is created by the [`iter`] method on a Spinoso [`String`]. See
-/// its documentation for more.
-///
-/// # Examples
-///
-/// ```
-/// # use spinoso_string::String;
-/// let s = String::utf8(b"Artichoke Ruby".to_vec());
-///
-/// let mut checksum: u32 = 0;
-/// for &byte in s.iter() {
-///     checksum += byte as u32;
-/// }
-/// assert_eq!(checksum, 1372);
-/// ```
-///
-/// [`String`]: crate::String
-/// [`iter`]: crate::String::iter
-#[derive(Debug, Clone)]
-pub struct Iter<'a>(slice::Iter<'a, u8>);
-
-impl<'a> Iter<'a> {
-    /// Views the underlying data as a subslice of the original data.
-    ///
-    /// This has the same lifetime as the original slice, and so the iterator
-    /// can continue to be used while this exists.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use spinoso_string::String;
-    /// let s = String::utf8(b"Artichoke Ruby".to_vec());
-    ///
-    /// // Then, we get the iterator:
-    /// let mut iter = s.iter();
-    /// assert_eq!(iter.as_slice(), b"Artichoke Ruby");
-    ///
-    /// // Next, we move to the second element of the slice:
-    /// iter.next();
-    /// // Now `as_slice` returns "rtichoke Ruby":
-    /// assert_eq!(iter.as_slice(), b"rtichoke Ruby");
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn as_slice(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-}
-
-/// Mutable [`String`] byte iterator.
-///
-/// This struct is created by the [`iter_mut`] method on a Spinoso [`String`].
-/// See its documentation for more.
-///
-/// # Examples
-///
-/// ```
-/// # use spinoso_string::String;
-/// let mut s = String::utf8(b"Artichoke Ruby".to_vec());
-///
-/// for byte in s.iter_mut() {
-///     *byte = b'z';
-/// }
-/// assert_eq!(s, "zzzzzzzzzzzzzz");
-/// ```
-///
-/// [`String`]: crate::String
-/// [`iter_mut`]: crate::String::iter_mut
-#[derive(Debug)]
-pub struct IterMut<'a>(slice::IterMut<'a, u8>);
-
-impl<'a> IterMut<'a> {
-    /// Views the underlying data as a subslice of the original data.
-    ///
-    /// To avoid creating &mut references that alias, this is forced to consume
-    /// the iterator.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use spinoso_string::String;
-    /// let mut s = String::utf8(b"Artichoke Ruby".to_vec());
-    /// {
-    ///     let mut iter = s.iter_mut();
-    ///     iter.next();
-    ///     assert_eq!(iter.into_slice(), b"rtichoke Ruby");
-    /// }
-    /// {
-    ///     let mut iter = s.iter_mut();
-    ///     *iter.next().unwrap() = b'a';
-    ///     *iter.nth(9).unwrap() = b'r';
-    /// }
-    /// assert_eq!(s, &b"artichoke ruby"[..]);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn into_slice(self) -> &'a mut [u8] {
-        self.0.into_slice()
-    }
-}
-
-/// An iterator that moves out of a string.
-///
-/// This struct is created by the `into_iter` method on `String` (provided by
-/// the [`IntoIterator`] trait).
-///
-/// # Examples
-///
-/// ```
-/// use spinoso_string::String;
-///
-/// let s = String::from("hello");
-/// let iter: spinoso_string::IntoIter = s.into_iter();
-/// ```
-#[derive(Debug)]
-pub struct IntoIter(vec::IntoIter<u8>);
-
-impl IntoIter {
-    /// Returns the remaining bytes of this iterator as a slice.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use spinoso_string::String;
-    ///
-    /// let s = String::from("abc");
-    /// let mut into_iter = s.into_iter();
-    /// assert_eq!(into_iter.as_slice(), &[b'a', b'b', b'c']);
-    /// let _ = into_iter.next().unwrap();
-    /// assert_eq!(into_iter.as_slice(), &[b'b', b'c']);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn as_slice(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-
-    /// Returns the remaining bytes of this iterator as a mutable slice.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use spinoso_string::String;
-    ///
-    /// let s = String::from("abc");
-    /// let mut into_iter = s.into_iter();
-    /// assert_eq!(into_iter.as_slice(), &[b'a', b'b', b'c']);
-    /// into_iter.as_mut_slice()[2] = b'z';
-    /// assert_eq!(into_iter.next(), Some(b'a'));
-    /// assert_eq!(into_iter.next(), Some(b'b'));
-    /// assert_eq!(into_iter.next(), Some(b'z'));
-    /// assert_eq!(into_iter.next(), None);
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn as_mut_slice(&mut self) -> &mut [u8] {
-        self.0.as_mut_slice()
-    }
-}
-
-/// Immutable [`String`] byte iterator.
-///
-/// This struct is created by the [`bytes`] method on a Spinoso [`String`]. See
-/// its documentation for more.
-///
-/// # Examples
-///
-/// ```
-/// # use spinoso_string::String;
-/// let s = String::utf8(b"Artichoke Ruby".to_vec());
-///
-/// let mut checksum: u32 = 0;
-/// for byte in s.bytes() {
-///     checksum += byte as u32;
-/// }
-/// assert_eq!(checksum, 1372);
-/// ```
-///
-/// [`String`]: crate::String
-/// [`bytes`]: crate::String::bytes
-#[derive(Debug, Clone)]
-pub struct Bytes<'a>(slice::Iter<'a, u8>);
-
-impl<'a> From<&'a [u8]> for Bytes<'a> {
-    #[inline]
-    fn from(bytes: &'a [u8]) -> Self {
-        Self(bytes.iter())
-    }
-}
-
-impl<'a> Bytes<'a> {
-    /// Views the underlying data as a subslice of the original data.
-    ///
-    /// This has the same lifetime as the original slice, and so the iterator
-    /// can continue to be used while this exists.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use spinoso_string::String;
-    /// let s = String::utf8(b"Artichoke Ruby".to_vec());
-    ///
-    /// // Then, we get the iterator:
-    /// let mut iter = s.bytes();
-    /// assert_eq!(iter.as_slice(), b"Artichoke Ruby");
-    ///
-    /// // Next, we move to the second element of the slice:
-    /// iter.next();
-    /// // Now `as_slice` returns "rtichoke Ruby":
-    /// assert_eq!(iter.as_slice(), b"rtichoke Ruby");
-    /// ```
-    #[inline]
-    #[must_use]
-    pub fn as_slice(&self) -> &[u8] {
-        self.0.as_slice()
-    }
-}
-
-/// Error returned when failing to construct a [`Center`] iterator.
-///
-/// This error is returned from [`String::center`]. See its documentation for
-/// more detail.
-///
-/// This error corresponds to the [Ruby `ArgumentError` Exception class].
-///
-/// When the **std** feature of `spinoso-string` is enabled, this struct
-/// implements [`std::error::Error`].
-///
-/// [Ruby `ArgumentError` Exception class]: https://ruby-doc.org/core-2.6.3/ArgumentError.html
-/// [`std::error::Error`]: https://doc.rust-lang.org/std/error/trait.Error.html
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
-pub enum CenterError {
-    /// Error returned when calling [`String::center`] with an empty padding
-    /// byte string.
-    ZeroWidthPadding,
-}
-
-impl CenterError {
-    pub const EXCEPTION_TYPE: &'static str = "ArgumentError";
-
-    /// Create a new zero width padding `CenterError`.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use spinoso_string::CenterError;
-    ///
-    /// const ERR: CenterError = CenterError::zero_width_padding();
-    /// assert_eq!(ERR.message(), "zero width padding");
-    /// ```
-    #[inline]
-    #[must_use]
-    pub const fn zero_width_padding() -> Self {
-        Self::ZeroWidthPadding
-    }
-
-    /// Retrieve the exception message associated with this center error.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// # use spinoso_string::CenterError;
-    /// let err = CenterError::zero_width_padding();
-    /// assert_eq!(err.message(), "zero width padding");
-    /// ```
-    #[inline]
-    #[must_use]
-    #[allow(clippy::unused_self)]
-    pub const fn message(self) -> &'static str {
-        "zero width padding"
-    }
-}
-
-impl fmt::Display for CenterError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let CenterError::ZeroWidthPadding = self;
-        f.write_str(self.message())
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for CenterError {}
-
-/// An iterator that yields a byte string centered within a padding byte string.
-///
-/// This struct is created by the [`center`] method on a Spinoso [`String`]. See
-/// its documentation for more.
-///
-/// # Examples
-///
-/// ```
-/// use spinoso_string::String;
-/// # fn example() -> Result<(), spinoso_string::CenterError> {
-/// let s = String::from("hello");
-///
-/// assert_eq!(s.center(4, None)?.collect::<Vec<_>>(), b"hello");
-/// assert_eq!(s.center(20, None)?.collect::<Vec<_>>(), b"       hello        ");
-/// assert_eq!(s.center(20, Some(&b"123"[..]))?.collect::<Vec<_>>(), b"1231231hello12312312");
-/// # Ok(())
-/// # }
-/// # example().unwrap();
-/// ```
-///
-/// This iterator is [encoding-aware]. [Conventionally UTF-8] strings are
-/// iterated by UTF-8 byte sequences.
-///
-/// ```
-/// use spinoso_string::String;
-/// # fn example() -> Result<(), spinoso_string::CenterError> {
-/// let s = String::from("💎");
-///
-/// assert_eq!(s.center(3, None)?.collect::<Vec<_>>(), " 💎 ".as_bytes());
-/// # Ok(())
-/// # }
-/// # example().unwrap();
-/// ```
-///
-/// [`center`]: crate::String::center
-/// [encoding-aware]: crate::Encoding
-/// [Conventionally UTF-8]: crate::Encoding::Utf8
-#[derive(Debug, Clone)]
-pub struct Center<'a, 'b> {
-    left: Take<Cycle<slice::Iter<'b, u8>>>,
-    next: Option<&'a [u8]>,
-    s: Chars<'a>,
-    right: Take<Cycle<slice::Iter<'b, u8>>>,
-}
-
-impl<'a, 'b> Default for Center<'a, 'b> {
-    #[inline]
-    fn default() -> Self {
-        Self::with_chars_width_and_padding(Chars::new(), 0, &[])
-    }
-}
-
-impl<'a, 'b> Center<'a, 'b> {
-    #[inline]
-    #[must_use]
-    pub(crate) fn with_chars_width_and_padding(s: Chars<'a>, padding_width: usize, padding: &'b [u8]) -> Self {
-        let pre_pad = padding_width / 2;
-        let post_pad = (padding_width + 1) / 2;
-        let left = padding.iter().cycle().take(pre_pad);
-        let right = padding.iter().cycle().take(post_pad);
-        Self {
-            left,
-            next: None,
-            s,
-            right,
-        }
-    }
-}
+pub use iter::{Bytes, Center, CenterError, IntoIter, Iter, IterMut};
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 enum CodePointRangeError {
@@ -1217,6 +864,34 @@ impl String {
     #[must_use]
     pub fn bytes(&self) -> Bytes<'_> {
         Bytes(self.buf.iter())
+    }
+}
+
+// Additional IntoIterator iterator
+impl IntoIterator for String {
+    type Item = u8;
+    type IntoIter = IntoIter;
+    /// Returns an iterator that moves over the remaining bytes of a slice
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use spinoso_string::String;
+    ///
+    /// let s = String::from("abc");
+    ///
+    /// let mut iterator = s.into_iter();
+    ///
+    /// assert_eq!(iterator.next(), Some(b'a'));
+    /// assert_eq!(iterator.next(), Some(b'b'));
+    /// assert_eq!(iterator.next(), Some(b'c'));
+    /// assert_eq!(iterator.next(), None);
+    /// ```
+
+    #[inline]
+    #[must_use]
+    fn into_iter(self) -> IntoIter {
+        IntoIter(self.buf.into_iter())
     }
 }
